Crate find_msvc_tools

These modules are all glue to support reading the MSVC version from the registry and from COM interfaces.

Structs

Tool

Tool found by find-msvc-tools

Enums

Env

VsVers

A version of Visual Studio

Traits

EnvGetter

Functions

find

Attempts to find a tool within an MSVC installation using the Windows registry as a point to search from.

find_tool

Similar to the find function above, this function will attempt the same operation (finding a MSVC tool in a local install) but instead returns a Tool which may be introspected.

find_tool_with_env

find_vs_version

Find the most recent installed version of Visual Studio

get_ucrt_dir

To find the Universal CRT we look in a specific registry key for where all the Universal CRTs are located and then sort them asciibetically to find the newest version. While this sort of sorting isn’t ideal, it is what vcvars does so that’s good enough for us.