find_msvc_tools

Function find_tool

pub fn find_tool(arch_or_target: &str, tool: &str) -> Option<Tool>

Similar to the find function above, this function will attempt the same operation (finding a MSVC tool in a local install) but instead returns a Tool which may be introspected.